+++ crates/engine/renderers/tataku-wgpu/src/shaders/gaussian_blur/buffer.rs
@@ -136,8 +136,8 @@ impl ReserveData<'_> {
 #[derive(bytemuck::Pod, bytemuck::Zeroable)]
 struct Blur2 {
     filter_size: u32,
-    x: f32,
-    y: f32,
-    width: f32,
-    height: f32,
+    x: u32,
+    y: u32,
+    width: u32,
+    height: u32,
 }
