+++ crates/engine/renderers/tataku-wgpu/src/shaders/gaussian_blur/data.rs
@@ -1,18 +1,18 @@
 #[derive(Copy, Clone, Debug, Default)]
 pub(crate) struct Params {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: u32,
+    pub y: u32,
+    pub width: u32,
+    pub height: u32,
 
     pub sigma: f32
 }
 impl Params {
     pub fn new(
-        x: f32,
-        y: f32,
-        width: f32,
-        height: f32,
+        x: u32,
+        y: u32,
+        width: u32,
+        height: u32,
         sigma: f32,
     ) -> Self {
         Self {
